pyproject.toml
[project]
name = "phildag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "cachetools==6.2.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "exceptiongroup==1.3.0",
    "google-auth==2.41.1",
    "google-genai==1.41.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.10",
    "pydantic-core==2.33.2",
    "pypdf2==3.0.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "websockets==15.0.1",
]
